Added support for queries with mapping into the library. This also includes new SWIG and a slightly changed CMakeLists.txt.

+++ src/API/DistanceQueryManagerWithMappingAPI.cpp
@@ -14,7 +14,7 @@ void DistanceQueryManagerWithMappingAPI::initializeCH(string chFile, string mapp
 }
 
 //______________________________________________________________________________________________________________________
-double DistanceQueryManagerWithMappingAPI::distanceQuery(unsigned int source, unsigned int target) {
+double DistanceQueryManagerWithMappingAPI::distanceQuery(long long unsigned int source, long long unsigned int target) {
     return qm -> findDistance(source, target);
 }
